Create README.md
This program is a simple calculator including addition, subtraction, multiplication, division, square, and square root which built using the Tkinter library in Python. It creates a window with an entry widget to display input and output and buttons to perform arithmetic operations and calculations. The calculator has buttons for digits 0-9, decimal point (.), arithmetic operators (+, -, *, /), square (^2), square root(sqrt), clear (C) and equals (=) button to evaluate the expression entered in the entry widget. When a button is clicked, the corresponding action is performed. If an error occurs during evaluation, it displays "Error" in the entry widget. Overall, the program provides a colourful graphical user interface for performing mathematical calculations.